requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchainhub
langchain-groq
faiss-cpu
sentence-transformers
